+++ visualization/app/codeCharta/state/store/appSettings/isAttributeSideBarVisible/isAttributeSideBarVisible.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/store"
+import { appSettingsSelector } from "../appSettings.selector"
+
+export const isAttributeSideBarVisibleSelector = createSelector([appSettingsSelector], appSettings => appSettings.isAttributeSideBarVisible)