+++ visualization/app/codeCharta/state/store/dynamicSettings/colorMetric/colorMetric.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { dynamicSettingsSelector } from "../dynamicSettings.selector"
+
+export const colorMetricSelector = createSelector([dynamicSettingsSelector], dynamicSettings => dynamicSettings.colorMetric)